requirements.txt
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
asyncpg==0.30.0
atlastk==0.13.5
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.4
django-celery-beat==2.8.1
django-timezone-field==7.1
et_xmlfile==2.0.0
fastapi==0.115.14
greenlet==3.2.3
h11==0.16.0
idna==3.10
iso8601==2.1.0
kombu==5.5.4
numpy==2.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
playwright==1.52.0
prompt_toolkit==3.0.51
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pypika-tortoise==0.6.1
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
starlette==0.46.2
tortoise==0.1.1
tortoise-orm==0.25.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.2
